serverName & serverVersion fit config
https://imgur.com/a/Vvnyakb
It get the serverName from the config. I think its only the first part of "This server is running NachoSpigot version" which will get customized.
It changes the serverVersion that the version don't look like "1.8.8", it look like "NachoSpigot-1.8.8".  Paper 1.17 has for example "git-Paper-114". 
The whole commit should only change the server log message & the timings. It doesnt affect the MOD in F3. If you want that it say everytime NachoSpigot in the log, you can simply change it back :P